// src/types/commonResponse.ts
import axios, { AxiosResponse } from 'axios';
/** Axios Response 데이터 형식
 *  config : 요청에 대한 axios 구성 설정
 *  data 서버가 제공한 응답 데이터
 *  headers : 헤더 정보
 *  request : 요청
 *  status : 응답 HTTP 상태 코드
 *  statusText : 응답 HTTP 상태 메시지
 */

// 본인 서버에서 내려주는 응답 구조
export default interface APIResponse<T> {
  status?: number;
  data?: any;
  statusCode?: number; // 상태코드 (보인 서버상태코드)
  errorCode?: number; // 에러코드 (본인 서버에러코드)
  message: string; // 메시지
  result: T; // 데이터 내용
  timestamp: Date; // 시간
}

interface uploadFile {
  video?: string[];
}

export interface APIResponseUpload<T> {
  statusText?: string;
  status?: number;
  data?: uploadFile;
  statusCode: number; // 상태코드 (보인 서버상태코드)
  errorCode: number; // 에러코드 (본인 서버에러코드)
  message: string; // 메시지
  result: T; // 데이터 내용
  timestamp: Date; // 시간
}